Define `sao` project structure.
This proposes a project layout for the `sao` package. The package code
is contained in `sao/` with corresponding test functions in `tests/`.
If needed, additional structure can be imposed by grouping related code
in subdirectories, as done for the current problem base class and the
example implementation `Rosenbrock`.

A `Makefile` and `setup.py` are provided in the root of the project. The
definitions in `setup.py` enable installing of the `sao` package, for
instance in a local virtual environment through `pip install -e .`. In
the Makefile basic functionality is provided to create such a local
virtual environment and install this package.

The tests of the code can be evaluated by manually invoking `pytest
tests/`. Additionally, a `tox.ini` file is provided to configure `Tox`
to automatically create virtual environments for testing and linting of
the code. Currently, for `tox` only a Python 3.9 environment is defined.