Update convergence with `&` and `|` operators.
This adds an implementation where the convergence criteria classes can
be combined with `&` and `|` to form new convergence criteria.
Additionally, this attempts to ensure that the variables are tracked
"externally", i.e. either through the `Storage` class (to be done) or by
passing the variables on initialisation of the convergence class.